chore: rename some of the shared helpers
