+++ packages/apps/human-app/frontend/src/modules/worker/components/jobs/evm-address.tsx
@@ -1,6 +1,6 @@
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
-import { shortenEscrowAddress } from '@/shared/helpers/address';
+import { shortenEscrowAddress } from '@/shared/helpers/evm';
 import { breakpoints } from '@/shared/styles/breakpoints';
 import { useIsMobile } from '@/shared/hooks/use-is-mobile';
 import { useColorMode } from '@/shared/hooks/use-color-mode';
