+++ packages/apps/human-app/frontend/src/shared/helpers/transfomers.ts
@@ -1,0 +1,5 @@
+import queryString from 'query-string';
+
+export function stringifyUrlQueryObject(obj: object) {
+  return queryString.stringify(obj);
+}
